working on constrained unit test, not yet working
