+++ batchglm/unit_test/glm_all/external.py
@@ -1,6 +1,4 @@
 from batchglm.unit_test.base_glm import Test_Accuracy_GLM, _Test_Accuracy_GLM_Estim
-from batchglm.unit_test.base_glm import Test_AccuracyAnalytic_GLM, _Test_AccuracyAnalytic_GLM_Estim
-from batchglm.unit_test.base_glm import Test_AccuracyConstrained_VGLM, _Test_AccuracyConstrained_VGLM_Estim
 from batchglm.unit_test.base_glm import Test_AccuracySizeFactors_GLM, _Test_AccuracySizeFactors_GLM_Estim
 from batchglm.unit_test.base_glm import Test_Graph_GLM, _Test_Graph_GLM_Estim
 from batchglm.unit_test.base_glm import Test_DataTypes_GLM, _Test_DataTypes_GLM_Estim
