+++ batchglm/unit_test/glm_all/test_acc_constrained_vglm_all.py
@@ -1,24 +1,18 @@
+import logging
+import numpy as np
 from typing import List
 import unittest
-import logging
-
-import numpy as np
 
 import batchglm.api as glm
 from batchglm.models.base_glm import _Estimator_GLM
 
-from .external import Test_AccuracyConstrained_VGLM, _Test_AccuracyConstrained_VGLM_Estim
 
-glm.setup_logging(verbosity="WARNING", stream="STDOUT")
-logger = logging.getLogger(__name__)
-
-
-class _Test_AccuracyConstrained_VGLM_ALL_Estim(_Test_AccuracyConstrained_VGLM_Estim):
+class _Test_AccuracyConstrained_VGLM_ALL_Estim:
 
     def __init__(
             self,
             simulator,
-            quick_scale,
+            train_scale,
             noise_model,
             constraints_loc,
             constraints_scale
@@ -33,7 +27,7 @@ class _Test_AccuracyConstrained_VGLM_ALL_Estim(_Test_AccuracyConstrained_VGLM_Es
             else:
                 raise ValueError("noise_model not recognized")
 
-        batch_size = 900
+        batch_size = 500
         provide_optimizers = {"gd": True, "adam": True, "adagrad": True, "rmsprop": True,
                               "nr": True, "nr_tr": True,
                               "irls": True, "irls_gd": True, "irls_tr": True, "irls_gd_tr": True}
@@ -41,7 +35,7 @@ class _Test_AccuracyConstrained_VGLM_ALL_Estim(_Test_AccuracyConstrained_VGLM_Es
         input_data = simulator.input_data
         design_loc = np.hstack([
             input_data.design_loc.values,
-            np.expand_dims(input_data.design_loc.values[:,0]-input_data.design_loc.values[:,-1], axis=-1)
+            np.expand_dims(input_data.design_loc.values[:, 0]-input_data.design_loc.values[:, -1], axis=-1)
         ])
         design_scale = design_loc.copy()
         input_data = InputData.new(
@@ -55,23 +49,165 @@ class _Test_AccuracyConstrained_VGLM_ALL_Estim(_Test_AccuracyConstrained_VGLM_Es
         estimator = Estimator(
             input_data=input_data,
             batch_size=batch_size,
-            quick_scale=quick_scale,
+            quick_scale=not train_scale,
             provide_optimizers=provide_optimizers,
             provide_batched=True,
-            init_a="standard",
-            init_b="standard"
-        )
-        super().__init__(
-            estimator=estimator,
-            simulator=simulator
+            provide_fim=True,
+            provide_hessian=True
         )
 
-class Test_AccuracyConstrained_VGLM_ALL(
-    Test_AccuracyConstrained_VGLM,
-    unittest.TestCase
-):
+        self.estimator = estimator
+        self.sim = simulator
+
+    def eval_estimation(
+            self,
+            estimator_store,
+            batched
+    ):
+        if batched:
+            threshold_dev_a = 0.5
+            threshold_dev_b = 0.5
+            threshold_std_a = 1
+            threshold_std_b = 20
+        else:
+            threshold_dev_a = 0.3
+            threshold_dev_b = 0.3
+            threshold_std_a = 1
+            threshold_std_b = 2
+
+        mean_dev_a = np.mean(estimator_store.a_var.values - self.sim.a_var.values)
+        std_dev_a = np.std(estimator_store.a_var.values - self.sim.a_var.values)
+        mean_dev_b = np.mean(estimator_store.b_var.values - self.sim.b_var.values)
+        std_dev_b = np.std(estimator_store.b_var.values - self.sim.b_var.values)
+
+        logging.getLogger("batchglm").info("mean_dev_a %f" % mean_dev_a)
+        logging.getLogger("batchglm").info("std_dev_a %f" % std_dev_a)
+        logging.getLogger("batchglm").info("mean_dev_b %f" % mean_dev_b)
+        logging.getLogger("batchglm").info("std_dev_b %f" % std_dev_b)
+
+        if np.abs(mean_dev_a) < threshold_dev_a and \
+                np.abs(mean_dev_b) < threshold_dev_b and \
+                std_dev_a < threshold_std_a and \
+                std_dev_b < threshold_std_b:
+            return True
+        else:
+            return False
+
+class Test_AccuracyConstrained_VGLM_ALL(unittest.TestCase):
     noise_model: str
     _estims: List[_Estimator_GLM]
+
+    def simulate(self):
+        self.simulate1()
+        self.simulate2()
+
+    def _simulate(self, num_batches):
+        sim = self.get_simulator()
+        sim.generate_sample_description(num_batches=num_batches, num_conditions=2)
+        sim.generate_params()
+        sim.size_factors = np.random.uniform(0.1, 2, size=sim.num_observations)
+        sim.generate_data()
+        logging.getLogger("batchglm").debug("Size factor standard deviation % f" %
+                                            np.std(sim.size_factors.data))
+        return sim
+
+    def simulate1(self):
+        self.sim1 = self._simulate(num_batches=2)
+
+    def simulate2(self):
+        self.sim2 = self._simulate(num_batches=0)
+
+    def simulator(self, train_loc):
+        if train_loc:
+            return self.sim1
+        else:
+            return self.sim2
+
+    def basic_test(
+            self,
+            batched,
+            train_loc,
+            train_scale
+    ):
+        if batched:
+            ts = ["IRLS_BATCHED"]
+        else:
+            ts = ["IRLS"]
+
+        # Encode equality constrained on overdetermined confounder coefficient.
+        if train_loc:
+            constraints = np.zeros([4, 3])
+            constraints[0, 0] = 1
+            constraints[1, 1] = 1
+            constraints[2, 2] = 1
+            constraints[3, 2] = -1
+        else:
+            constraints = np.zeros([3, 2])
+            constraints[0, 0] = 1
+            constraints[1, 1] = 1
+            constraints[2, 1] = -1
+
+        for x in ts:
+            logging.getLogger("batchglm").debug("algorithm: %s" % x)
+            estimator = _Test_AccuracyConstrained_VGLM_ALL_Estim(
+                simulator=self.simulator(train_loc=train_loc),
+                train_scale=train_scale,
+                noise_model=self.noise_model,
+                constraints_loc=constraints,
+                constraints_scale=constraints,
+            )
+            estimator.estimator.initialize()
+            estimator.estimator.train_sequence(training_strategy=ts)
+            estimator_store = estimator.estimator.finalize()
+            success = estimator.eval_estimation(
+                estimator_store=estimator_store,
+                batched=batched
+            )
+            assert success, "%s did not yield exact results" % x
+
+        return True
+
+    def _test_full_a_and_b(self):
+        return self.basic_test(
+            batched=False,
+            train_loc=True,
+            train_scale=True
+        )
+
+    def _test_full_a_only(self):
+        return self.basic_test(
+            batched=False,
+            train_loc=True,
+            train_scale=False
+        )
+
+    def _test_full_b_only(self):
+        return self.basic_test(
+            batched=False,
+            train_loc=False,
+            train_scale=True
+        )
+
+    def _test_batched_a_and_b(self):
+        return self.basic_test(
+            batched=True,
+            train_loc=True,
+            train_scale=True
+        )
+
+    def _test_batched_a_only(self):
+        return self.basic_test(
+            batched=True,
+            train_loc=True,
+            train_scale=False
+        )
+
+    def _test_batched_b_only(self):
+        return self.basic_test(
+            batched=True,
+            train_loc=False,
+            train_scale=True
+        )
 
     def get_simulator(self):
         if self.noise_model is None:
@@ -86,52 +222,20 @@ class Test_AccuracyConstrained_VGLM_ALL(
 
         return Simulator(num_observations=10000, num_features=10)
 
-    def basic_test(
-            self,
-            batched,
-            train_loc,
-            train_scale
-    ):
-        algos = ["ADAM", "NR_TR", "IRLS_GD_TR"]
-        # Encode equality constrained on overdetermined confounder coefficient.
-        if train_loc:
-            constraints = np.zeros([4, 3])
-            constraints[0, 0] = 1
-            constraints[1, 1] = 1
-            constraints[2, 2] = 1
-            constraints[3, 2] = -1
-        else:
-            constraints = np.zeros([3, 2])
-            constraints[0, 0] = 1
-            constraints[1, 1] = 1
-            constraints[2, 1] = -1
-
-        estimator = _Test_AccuracyConstrained_VGLM_ALL_Estim(
-            simulator=self.simulator(train_loc=train_loc),
-            quick_scale=False if train_scale else True,
-            noise_model=self.noise_model,
-            constraints_loc=constraints,
-            constraints_scale=constraints,
-        )
-        return self._basic_test(
-            estimator=estimator,
-            batched=batched,
-            algos=algos
-        )
 
     def _test_full(self):
         self.simulate()
-        logger.debug("* Running tests for full data")
-        super()._test_full_a_and_b()
-        super()._test_full_a_only()
-        super()._test_full_b_only()
+        logging.getLogger("batchglm").debug("* Running tests for full data")
+        self._test_full_a_and_b()
+        self._test_full_a_only()
+        self._test_full_b_only()
 
     def _test_batched(self):
         self.simulate()
-        logger.debug("* Running tests for batched data")
-        super()._test_batched_a_and_b()
-        super()._test_batched_a_only()
-        super()._test_batched_b_only()
+        logging.getLogger("batchglm").debug("* Running tests for batched data")
+        self._test_batched_a_and_b()
+        self._test_batched_a_only()
+        self._test_batched_b_only()
 
 
 class Test_AccuracyConstrained_VGLM_NB(
@@ -145,7 +249,7 @@ class Test_AccuracyConstrained_VGLM_NB(
     def test_full_nb(self):
         logging.getLogger("tensorflow").setLevel(logging.ERROR)
         logging.getLogger("batchglm").setLevel(logging.WARNING)
-        logger.error("Test_AccuracyConstrained_VGLM_NB.test_full_nb()")
+        logging.getLogger("batchglm").error("Test_AccuracyConstrained_VGLM_NB.test_full_nb()")
 
         self.noise_model = "nb"
         self._test_full()
@@ -153,7 +257,7 @@ class Test_AccuracyConstrained_VGLM_NB(
     def test_batched_nb(self):
         logging.getLogger("tensorflow").setLevel(logging.ERROR)
         logging.getLogger("batchglm").setLevel(logging.WARNING)
-        logger.error("Test_AccuracyConstrained_VGLM_NB.test_batched_nb()")
+        logging.getLogger("batchglm").error("Test_AccuracyConstrained_VGLM_NB.test_batched_nb()")
 
         self.noise_model = "nb"
         self._test_batched()
@@ -169,7 +273,7 @@ class Test_AccuracyConstrained_VGLM_NORM(
     def test_full_norm(self):
         logging.getLogger("tensorflow").setLevel(logging.ERROR)
         logging.getLogger("batchglm").setLevel(logging.WARNING)
-        logger.error("Test_AccuracyConstrained_VGLM_NORM.test_full_norm()")
+        logging.getLogger("batchglm").error("Test_AccuracyConstrained_VGLM_NORM.test_full_norm()")
 
         self.noise_model = "norm"
         self._test_full()
@@ -177,7 +281,7 @@ class Test_AccuracyConstrained_VGLM_NORM(
     def test_batched_norm(self):
         logging.getLogger("tensorflow").setLevel(logging.ERROR)
         logging.getLogger("batchglm").setLevel(logging.WARNING)
-        logger.error("Test_AccuracyConstrained_VGLM_NORM.test_batched_norm()")
+        logging.getLogger("batchglm").error("Test_AccuracyConstrained_VGLM_NORM.test_batched_norm()")
 
         self.noise_model = "norm"
         self._test_batched()
